### 2024-04-29 #6694 @editorialbot
[PRE REVIEW]: DTW-C++: Fast dynamic time warping and clustering of time series data
**Submitting author:** @ElektrikAkar (<a href="http://orcid.org/0000-0003-2392-9771">Volkan Kumtepeli</a>)
**Repository:** https://github.com/Battery-Intelligence-Lab/dtw-cpp
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
