### 2024-04-29 #6694 @crvernon
:wave: @ElektrikAkar - thanks for your submission to JOSS. We are currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
