### 2024-05-27 #6694 @logological
@TahiriNadia, @lmcinnes, @taqtiqa-mark, @breznak would any of you be willing and able to review this submission for JOSS?
