### 2024-05-30 #6694 @logological
@i64 Thanks for volunteering.  In order for me to assess your suitability as a reviewer, would you mind providing some background on your expertise in the field of software and in the field of this submission?  I may have to double-check on this, but I might also need to know your real-life name or identity, if only to exclude the possibility of there being a conflict of interest.  You could provide all this information here or by private e-mail to me at tristan@logological.org.
