### 2024-06-04 #6694 @logological
@ghammad, @mikeaalv, @tomelse, @sgeorge91 would any of you be willing and able to review this submission for JOSS?
