### 2024-06-12 #6694 @logological
@ZhenchenHong, @SimonMolinsky, @marcoapintoo, @ulf1 would any of you be willing and able to review this submission for JOSS?
