### 2024-06-12 #6881 @editorialbot
[REVIEW]: DTW-C++: Fast dynamic time warping and clustering of time series data
**Submitting author:** @ElektrikAkar (<a href="http://orcid.org/0000-0003-2392-9771">Volkan Kumtepeli</a>)
**Repository:** https://github.com/Battery-Intelligence-Lab/dtw-cpp
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
