### 2024-07-05 Battery-Intelligence-Lab/DTWpp#12 @ZhenchenHong
Minor comments in executive file location 
Hi authors, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/6881#issue-2349216956. Could you please add the directory of `dtwc_main` to help the user to specify the location? To be specifically, in `https://battery-intelligence-lab.github.io/dtw-cpp/1_getting_started/direct_use.html`, you may add the `./dtw-cpp/bin` to direct people to the correct path. Thanks!
