### 2024-07-12 4e3b51e
Update direct_use.md to include bin folder
The `dtw-cpp/bin` folder for `dtwc_main` executable is denoted in the instructions.
