+++ docs/1_getting_started/direct_use.md
@@ -12,4 +12,4 @@ Although DTW-C++ is easiest to use via the command line interface executable, it
 - Download the source code and extract it if it is zipped.
 - Copy your data into the `data` folder in one of the [supported data formats](supported_data.md).
 - Edit `main.cpp` in the `dtwc` folder using the examples in the `examples` folder or [examples page](examples.md).
-- Compile the code using [these instructions](run.md), and use the `dtwc_main` executable after compilation. 
+- Compile the code using [these instructions](run.md), and after compilation use the `dtwc_main` executable located in the `./dtw-cpp/bin` folder.
