### 2024-07-12 Battery-Intelligence-Lab/DTWpp#12 @ElektrikAkar
Hi, thank you very much for your comment, I have added the folder in the instructions in commit 4e3b51ef5b6bce9ccbc6690cb755bfbb0350d548. Please let me know if any further clarification is needed.
