### 2024-07-12 Battery-Intelligence-Lab/DTWpp#13 @i64
Benchmarking dependencies and references
Hi authors, I am reviewing the code from openjournals/joss-reviews#6881

- No "requirments.txt" (or something equivalent) is provided to run the python benchmarks; however, they use third-party libraries.
- In the [bibliography](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/4e3b51ef5b6bce9ccbc6690cb755bfbb0350d548/joss/paper.bib#L46C18-L46C30), the library "dtaidistance" ([github](https://github.com/wannesm/dtaidistance/releases), [zenodo](https://zenodo.org/records/5901139)) references an old version from 2020 (v2.0.0), while there is a more modern version from 2022(v2.3.10). Is the 2020 version used mistakenly?
