### 2024-07-12 Battery-Intelligence-Lab/DTWpp#14 @i64
Reference duplication
Hi authors, I am reviewing the code from [openjournals/joss-reviews#6881](https://github.com/openjournals/joss-reviews/issues/6881#issue-2349216956).

The reference "The UCR Time Series Classification Archive" is duplicated [[1](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/4e3b51ef5b6bce9ccbc6690cb755bfbb0350d548/joss/paper.bib#L114-L120), [2](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/4e3b51ef5b6bce9ccbc6690cb755bfbb0350d548/joss/paper.bib#L65-L73)] in the bibliography.
