### 2024-07-12 #6881 @i64
@logological There are two "Statement of need" items in the checklist. In the paper, they have it as a section; however, in the documentation what should be our expectation? Can introduction/about sections of the documentation considered to be enough?
