### 2024-07-12 Battery-Intelligence-Lab/DTWpp#15 @i64
Community guideline improvement
Hi authors, I am reviewing the code from [openjournals/joss-reviews#6881](https://github.com/openjournals/joss-reviews/issues/6881#issue-2349216956).

The "develop" folder contains community guidelines. It could improved to better cover the second item in "Community Guidlines" in the JOSS' checklist.
