### 2024-07-12 Battery-Intelligence-Lab/DTWpp#16 @i64
Third-party software installation guidelines
Hi authors, I am reviewing the code from [openjournals/joss-reviews#6881](https://github.com/openjournals/joss-reviews/issues/6881#issue-2349216956).

The third-party software (e.g., CMake, OpenMP, Gurobi) installation guidelines are missing and could be integrated into the documentation.
