### 2024-07-12 Battery-Intelligence-Lab/DTWpp#12 @i64
I agree with @ZhenchenHong. I think the authors can also fix location confusion in the "running" section.
