### 2024-07-12 Battery-Intelligence-Lab/DTWpp#16 @i64
Another thing is, in the Linux section, it should be made clear which linux distribution it is. Then for example the installation comments could be such as

```
sudo apt install build-essential
sudo apt install cmake libomp-dev
```
