### 2024-07-12 Battery-Intelligence-Lab/DTWpp#17 @i64
Installation documentation 
Hi authors, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/6881.

In installation
`cd DTWC++  is wrong it should be cd dtw-cpp`

running is itself the installation, and overview, installation, running should be in order. Also -j32 should not be assumed where you could use this command in in linux to determine number of processors.

`cmake --build . -j$(( $(nproc) * 2 -1))`
