### 2024-07-12 Battery-Intelligence-Lab/DTWpp#18 @i64
Memory profiler overhead in benchmarks
Hi authors, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/6881

For other libraries, benchmark code includes a memory profiler (trace malloc), while the authors' library does not include a memory profiler.

When obtaining benchmarks, is the overhead caused by the memory profiler considered?
