### 2024-07-17 a9d6168
removed tracemalloc
Removed the tracemalloc memory tracker to check the timing benchmarking results are not significantly influenced
