### 2024-07-17 Battery-Intelligence-Lab/DTWpp#18 @beckyperriment
Hi, thank you for bringing this to our attention. We have partially rerun the benchmarks for DTAIDistance without the memory profiler to check this (commit #[a9d6168](https://github.com/Battery-Intelligence-Lab/dtw-cpp/commit/a9d6168b9c8729688ce3700c28cbd32063679689) ), and there is no significant speed increase when it is removed. There is the expected variation in the smaller clustering problems, which are solved in<10 seconds. For a couple larger clustering problems, as examples:

1. Chlorine concentration
      - DTW-C++ = 70 seconds
      - DTAIDistance with tracemalloc = 201 seconds
      - DTAIDistance without tracmalloc =182 seconds

2. cinCECGTorse
      - DTW-C++ = 1104 seconds
      - DTAIDistance with tracemalloc = 1955 seconds
      - DTAIDistance  without tracmalloc =2079 seconds

We believe this satisfies concerns over the impact of the memory tracker on benchmarking, but if you would like further clarification, please let us know!
