### 2024-07-30 4750f58
JOSS paper minor update
