+++ joss/paper.bib
@@ -42,11 +42,13 @@
    year = {2011},
 }
 
-@article{meert2020wannesm,
-  title={wannesm/dtaidistance v2. 0.0},
-  author={Meert, Wannes and Hendrickx, Kilian and Van Craenendonck, Toon},
+@article{meert2022wannesm,
+  title={wannesm/dtaidistance v2.3.10},
+  author={Meert, Wannes and Hendrickx, Kilian and Van Craenendonck, Toon and Robberechts, Pieter and Blockeel, Hendrik and Davis, Jesse},
   journal={Zenodo},
-  year={2020}
+  year={2022}, 
+  doi={10.5281/zenodo.1202378},
+  url={https://zenodo.org/records/7158824}
 }
 
 @inproceedings{Shahapure2020,
@@ -60,6 +62,18 @@
    publisher = {Institute of Electrical and Electronics Engineers Inc.},
    title = {Cluster quality analysis using silhouette score},
    year = {2020},
+}
+
+@article{dau2019ucr,
+  title={The UCR time series archive},
+  author={Dau, Hoang Anh and Bagnall, Anthony and Kamgar, Kaveh and Yeh, Chin-Chia Michael and Zhu, Yan and Gharghabi, Shaghayegh and Ratanamahatana, Chotirat Ann and Keogh, Eamonn},
+  journal={IEEE/CAA Journal of Automatica Sinica},
+  volume={6},
+  number={6},
+  pages={1293--1305},
+  year={2019},
+  publisher={IEEE},
+  doi={10.1109/JAS.2019.1911747}
 }
 
 @misc{UCRArchive2018,
@@ -111,14 +125,6 @@ note = {\url{https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}}
    year = {2020},
 }
 
-@misc{Dau2018,
-   author = {Hoang Anh Dau and Eamonn Keogh and Kaveh Kamgar and Chin-Chia Michael Yeh and Yan Zhu and Shaghayegh Gharghabi and Chotirat Ann Ratanamahatana and Yanping and Bing Hu and Nurjahan Begum and Anthony Bagnall and Abdullah Mueen and Gustavo Batista},
-   month = {10},
-   title = {The UCR Time Series Classification Archive},
-   url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/},
-   year = {2018},
-}
-
 @article{Huangfu2018,
    abstract = {This paper introduces the design and implementation of two parallel dual simplex solvers for general large scale sparse linear programming problems. One approach, called PAMI, extends a relatively unknown pivoting strategy called suboptimization and exploits parallelism across multiple iterations. The other, called SIP, exploits purely single iteration parallelism by overlapping computational components when possible. Computational results show that the performance of PAMI is superior to that of the leading open-source simplex solver, and that SIP complements PAMI in achieving speedup when PAMI results in slowdown. One of the authors has implemented the techniques underlying PAMI within the FICO Xpress simplex solver and this paper presents computational results demonstrating their value. In developing the first parallel revised simplex solver of general utility, this work represents a significant achievement in computational optimization.},
    author = {Q. Huangfu and J. A.J. Hall},
@@ -138,7 +144,7 @@ note = {\url{https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}}
 @misc{gurobi,
   author = {{Gurobi Optimization, LLC}},
   title = {{Gurobi Optimizer Reference Manual}},
-  year = 2023,
+  year = 2024,
   url = "https://www.gurobi.com"
 }
 
[…]
