+++ benchmark/requirements.txt
@@ -1,0 +1,4 @@
+pandas==1.3.3
+glob2==0.7
+dtaidistance==2.3.10
+numpy==1.21.2
