+++ joss/paper.bib
@@ -154,7 +154,7 @@ note = {\url{https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}}
       year={2023},
       eprint={2307.04904},
       archivePrefix={arXiv},
-      primaryClass={eess.SP}
+      primaryClass={eess.SP},
       doi = {10.48550/arXiv.2307.04904}
 }
 
