+++ joss/paper.bib
@@ -83,7 +83,7 @@ author = {Dau, Hoang Anh and Keogh, Eamonn and Kamgar, Kaveh and Yeh, Chin-Chia 
 	  and Begum, Nurjahan and Bagnall, Anthony and Mueen, Abdullah and Batista, Gustavo, and Hexagon-ML},
 year = {2018},
 month = {October},
-note = {\url{https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}}
+url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
 }
 
 @article{Sakoe1978,
