### 2024-07-30 Battery-Intelligence-Lab/DTWpp#13 @ElektrikAkar
Hi i64,

Thank you very much for your comments!

- Yes, we have now added a [requirements.txt](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/main/benchmark/requirements.txt) for benchmarks. However, I should note that the `TSlearn` and `dtaidistance` libraries require different `numpy` versions. Therefore, we wrote the [requirements.txt](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/main/benchmark/requirements.txt) considering `dtaidistance` as we only presented results regarding this library in the paper.
- You are right that we referred to an older version of the `dtaidistance` library. I think we copied an older URL from the repository of the software. Checking the version we used, we see that it was version `2.3.10`; therefore, we updated the reference accordingly (which is also reflected in the [requirements.txt](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/main/benchmark/requirements.txt)).

See 4750f5837be4a97807b2e69cf92e98f66887fcfe for the details.
