### 2024-07-30 Battery-Intelligence-Lab/DTWpp#18 @ElektrikAkar
In addition to Becky's comment/changes, I have added a [remark](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/c44f1d4fb4f6e0def7160b12f58225a46f05a001/joss/paper.md?plain=1#L166) to the table indicating that the results were obtained in the presence of `tracemalloc`, which _may_ slightly affect the results. However, I also did not see any significant effect of `tracemalloc`, probably because we did not collect multiple snapshots.

See 4750f5837be4a97807b2e69cf92e98f66887fcfe for the details.
