### 2024-07-30 Battery-Intelligence-Lab/DTWpp#14 @ElektrikAkar
Hi i64,

Thank you! We changed one of the references with the [IEEE paper](https://doi.org/10.1109/JAS.2019.1911747) and other with the [data repository](https://www.cs.ucr.edu/~eamonn/time_series_data_2018/).

See 4750f5837be4a97807b2e69cf92e98f66887fcfe for the details.
