### 2024-07-30 Battery-Intelligence-Lab/DTWpp#14 @i64
Hi @ElektrikAkar,

I checked the commit. LGTM!
