### 2024-07-30 Battery-Intelligence-Lab/DTWpp#18 @i64
Hi @ElektrikAkar and @beckyperriment,

Thank you for adding the remark and calculating the tracemallocs overhead
