### 2024-07-30 Battery-Intelligence-Lab/DTWpp#13 @i64
Hi @ElektrikAkar,

Thank you for the update and  the commit. LGTM
