### 2024-07-31 c9244b6
Update contributing guidelines
