+++ docs/7_develop/index.md
@@ -1,0 +1,15 @@
+---
+layout: default
+title: Contributing
+nav_order: 1
+---
+
+Here you may find the documentation how to contribute!
+
+[Contributing guideline](CONTRIBUTING.md)
+
+[Coding conventions](conventions.md)
+
+[The list of contributors](contributors.md)
+
+[TODO list](https://github.com/Battery-Intelligence-Lab/dtw-cpp/tree/main/develop/TODO.md)
