+++ develop/contributors.md
@@ -1,26 +1,37 @@
+---
+layout: default
+title: Contributors
+nav_order: 4
+---
+
 # List of contributors 
-Please feel free to add your name into the list! 
+
+If you PR request involving more than a one-line change is approved, your name will be added to the contributor list referring to the regarding PRs.
+
+## Core contributors:
+
+**Rebecca Perriment:** project owner
+- uses code to cluster data,
+- reviews literature and generates novel ideas,
+- documents the code, 
+- created benchmark studies,
+- wrote JOSS paper and improved upon reviews.
 
 
-## Core contributors
-
-
-**Rebecca Perriment:** project owner.
-- using code to cluster data, 
-- review literature and come up with novel ideas.
-- documented the code. 
-
-
-**Volkan Kumtepeli:** main developer. 
-- wrote the initial code.
-- updates the code according to the requirements. 
-- built the documentation website.
-- organised file and code structure.
+**Volkan Kumtepeli:** main developer
+- developed the initial code,
+- updates the code according to requirements,
+- built the documentation website,
+- organised file and code structure,
+- improved the code upon JOSS paper reviews, 
+- reviewed documentation and the paper. 
 
 
 **David A. Howey:** project administrator.
-- contributed via project administration, supervison, documentation, review and editing. 
+- contributes via project administration, supervison, documentation, review, and editing. 
 
 
 ## Other contributors
+
+We welcome contributors from the community.
 
