+++ CONTRIBUTING.md
@@ -1,1 +1,43 @@
-Please see [develop folder](https://github.com/Battery-Intelligence-Lab/dtw-cpp/tree/main/develop) for detailed instructions for prospective contributors. Many thanks in advance! 
+---
+layout: default
+title: Guideline
+nav_order: 2
+---
+
+# Contributing
+
+DTW-C++ is an open-source project open to both academic and commercial applications. We use GitHub issues to track all bugs and features. We welcome and appreciate any contributions from the community. For our code of conduct, please refer to the [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/).
+
+## Opening issues
+
+We use GitHub issues to track all bugs and features. Feel free to open an issue if you encounter any problems or have suggestions for improvements. If an issue has been closed but you believe it still needs attention, please re-open it.
+
+## Pull requests (PRs)
+
+If you have a solution to an issue, we welcome [GitHub pull requests](https://help.github.com/articles/creating-a-pull-request/) from the community.
+
+1. Open an issue describing the problem your PR will address, if one does not already exist. Outline your approach to solving it.
+2. We do not have a contributor license agreement yet; but please just agree that your contributions will be publicly open. 
+3. Fork the repository to your own GitHub account.
+4. Code your solution. Please do not forget to read the [coding conventions](develop/conventions.md)!
+5. Squash your commits into a single commit with a clear message describing the PR.
+6. Include tests that verify the changes introduced by your PR. Ensure that your tests cover the bug fixes or new features thoroughly.
+7. Submit your PR, referencing the related issue.
+8. Provide clear steps to reproduce any bugs your PR addresses.
+
+The core team will review your PR as soon as possible. If accepted, we will add you to the [list of contributors](develop/contributors.md).
+
+For further details, refer to the [develop folder](https://github.com/Battery-Intelligence-Lab/dtw-cpp/tree/main/develop). Thank you for your contributions!
+
+### How to find good first issues to work on
+
+Issues labelled with `good first issue` are great starting points. Additionally, check our [TODO.md](https://github.com/Battery-Intelligence-Lab/dtw-cpp/tree/main/develop/TODO.md) for next steps that might not be listed as issues yet.
+
+## Learning Material
+
+### Introduction to Git and GitHub
+
+- Ensure you have a [GitHub account](https://github.com/signup/free).
+- Learn Git:
+  - GitHub Help: [Good Resources for Learning Git and GitHub][good-git-resources]
+- [GitHub Flow Guide](https://guides.github.com/introduction/flow/): Step-by-step instructions for GitHub Flow
