### 2024-07-31 Battery-Intelligence-Lab/DTWpp#15 @ElektrikAkar
Hi i64,

Thank you very much for your comment. We have updated our [community guidelines](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/main/CONTRIBUTING.md) along with our [conventions](https://github.com/Battery-Intelligence-Lab/dtw-cpp/blob/main/develop/conventions.md).

We have also added these guidelines to the [latest section](https://battery-intelligence-lab.github.io/dtw-cpp/7_develop/CONTRIBUTING.html) of our documentation website, making it easier for prospective contributors to find them.

Please see commit c9244b665df8af12c4301aec296da5ee8f10c7d4 for more details.
