### 2024-08-01 Battery-Intelligence-Lab/DTWpp#15 @i64
Hi @ElektrikAkar,

I checked the commit. LGTM!
