+++ dtwc/parallelisation.hpp
@@ -2,7 +2,7 @@
  * @file parallelisation.hpp
  * @brief Header for parallelisation functions.
  *
- * @details This header file provides functionalities for parallelising tasks using OpenMP.
+ * @details This header file provides functionalities for parallelising tasks using standard parallelisation.
  * It includes functions for running individual tasks in parallel and adjusting
  * the level of parallelism. Functions are templated to support various task types.
  *
@@ -16,14 +16,16 @@
 #include "types/Range.hpp"
 
 #include <cstddef>
-#include <omp.h>
+
+#include <execution>
+#include <algorithm>
 
 namespace dtwc {
 
 /**
- * @brief Runs a given task in parallel using OpenMP.
+ * @brief Runs a given task in parallel using std:;execution.
  *
- * This function executes the provided task in parallel, leveraging OpenMP's dynamic scheduling.
+ * This function executes the provided task in parallel, leveraging std:execution::par_unseq's dynamic scheduling.
  * The dynamic scheduling is advantageous for tasks with varying completion times. It allows for
  * better load balancing across threads.
  *
@@ -33,15 +35,13 @@ namespace dtwc {
  * @param isParallel Flag to enable/disable parallel execution (default is true).
  */
 template <typename Tfun>
-void run_openmp(Tfun &task_indv, size_t i_end, bool isParallel = true)
+void run_std(Tfun &task_indv, size_t i_end, bool isParallel = true)
 {
-  if (isParallel) {
-#pragma omp parallel for schedule(dynamic) // As some take less time static scheduling is 2x slower.
-    for (int i = 0; i < i_end; i++)
-      task_indv(i);
-  } else
-    for (int i = 0; i < i_end; i++)
-      task_indv(i);
+  auto range = Range(i_end);
+  if (isParallel)
+    std::for_each(std::execution::par_unseq, range.begin(), range.end(), task_indv);
+  else
+    std::for_each(std::execution::seq, range.begin(), range.end(), task_indv);
 }
 
 /**
@@ -49,7 +49,7 @@ void run_openmp(Tfun &task_indv, size_t i_end, bool isParallel = true)
  *
  * @details This function provides a higher level of control for parallel task execution.
  * It decides whether to use parallelism based on the provided maximum number of parallel workers.
- * It delegates the task execution to 'run_openmp'.
+ * It delegates the task execution to 'run_std'.
  *
  * @tparam Tfun The type of the task function.
  * @param task_indv Reference to the task function to be executed.
@@ -59,6 +59,6 @@ void run_openmp(Tfun &task_indv, size_t i_end, bool isParallel = true)
 template <typename Tfun>
 void run(Tfun &task_indv, size_t i_end, size_t numMaxParallelWorkers = 32)
 {
-  run_openmp(task_indv, i_end, numMaxParallelWorkers != 1);
+  run_std(task_indv, i_end, numMaxParallelWorkers != 1);
 }
 } // namespace dtwc
