+++ dtwc/CMakeLists.txt
@@ -36,8 +36,3 @@ target_link_libraries(dtwc++
   PUBLIC 
   armadillo
 )
-
-find_package(OpenMP)
-if(OpenMP_CXX_FOUND)
-  target_link_libraries(dtwc++ PUBLIC OpenMP::OpenMP_CXX)
-endif()

[…]
