### 2024-08-15 f93367f
target_compile_definitions for root folder
