### 2024-08-15 cb37921
CURRENT_ROOT_FOLDER is added
