+++ dtwc/CMakeLists.txt
@@ -2,6 +2,7 @@ cmake_minimum_required(VERSION 3.21)
 
 add_library(dtwc++ STATIC)
 target_compile_definitions(dtwc++ PUBLIC DTWC_ROOT_FOLDER="${DTWC_ROOT_FOLDER}")
+target_compile_definitions(dtwc++ PUBLIC CURRENT_ROOT_FOLDER="${CURRENT_ROOT_FOLDER}")
 
 add_subdirectory(mip)
 
