### 2024-08-16 576ab3e
Create example project.
