+++ docs/1_getting_started/direct_use.md
@@ -1,15 +1,0 @@
----
-layout: default
-title: Direct use
-nav_order: 4
----
-
-# Direct use
-
-Although DTW-C++ is easiest to use via the command line interface executable, it can also be used directly as follows:
-
-- Install [the dependencies](dependencies.md).
-- Download the source code and extract it if it is zipped.
-- Copy your data into the `data` folder in one of the [supported data formats](supported_data.md).
-- Edit `main.cpp` in the `dtwc` folder using the examples in the `examples` folder or [examples page](examples.md).
-- Compile the code using [these instructions](1_installation.md), and after compilation use the `dtwc_main` executable located in the `./dtw-cpp/bin` folder.
