+++ docs/1_getting_started/1_installation.md
@@ -14,64 +14,164 @@ DTW-C++ does not offer any binaries or wrappers in other languages at the moment
 
 # Dependencies
 
-DTW-C++ aims to 
+DTW-C++ aims to be easily compilable and usable; therefore, it includes only a few libraries where most of the dependencies are automatically installed.  
+
+There are several pre-requisite installations required to compile and run DTW-C++.
+
+The following dependencies need to be manually installed by the user if they do not already exist:
+- [CMake](https://cmake.org/)
+- A suitable compiler (Clang, GCC, MSVC, etc.)
+- Gurobi (optional, if not installed then HiGHS will be used as the MIP solver)
+
+The following dependencies are installed by the CPM package manager: 
+- [HiGHS](https://highs.dev/) as an open source MIP solver alternative to Gurobi. 
+- [CLI11](https://github.com/CLIUtils/CLI11) (for command line interface)
+- [Catch2](https://github.com/catchorg/Catch2/) (for testing)
+- [Armadillo](https://arma.sourceforge.net/) (for matrix reading)
+
+## CMake and compilers
+
+[CMake](https://cmake.org/) is a metabuild system required to provide the correct compilation commands to the compiler. It will be explained how to install CMake and compilers depending on your operating system below. 
+
+## Gurobi
+
+Gurobi is a powerful optimisation solver that is free for academic use. If you do not wish to use Gurobi, HiGHS will be used instead. Please see the following guidelines for installation on [Ubuntu](https://www.youtube.com/watch?v=yNmeG6Wom1o), [macOS](https://www.youtube.com/watch?v=ZcL-NmckTxQ), [Windows](https://www.youtube.com/watch?v=z7t0p5J9YcQ), and [further information](https://support.gurobi.com/hc/en-us/sections/360010017231-Platforms-and-Installation)
 
 
+# Building from the source
 
+DTW-C++ aims to be compatible with different compilers and platforms. You may easily install DTW-C++ using CMake (although it is not an absolute requirement). Therefore, you need a suitable C++ compiler (preferably [GCC](https://gcc.gnu.org/)) and [CMake](https://cmake.org/) to follow this installation guide.   
 
+## Linux (Debian / Ubuntu 20.04+)
 
-## Building from the source
+Here we present the default compilation comments targetting new Ubuntu versions above 20.04. As long as there is `CMake 3.21` and a `C++17` capable compiler is installed DTW-C++ should work. However, the compilers installed with default commands in older Ubuntu versions may be older compilers that do not support some of the functionalities in this code directly. Therefore, please refer to [install a newer version of GCC](https://linuxize.com/post/how-to-install-gcc-compiler-on-ubuntu-18-04/) for Ubuntu versions 18.04 or below.
 
-DTW-C++ aims to be compatible with different compilers and platforms. 
+1. Install the essential libraries for building the project
 
+    ```bash
+    sudo apt update
+    sudo apt install -y build-essential cmake-extras cmake
+    ```
 
-You may easily install DTW-C++ using CMake (although it is not an absolute requirement). Therefore, you need a suitable C++ compiler (preferably [GCC](https://gcc.gnu.org/)) and [CMake](https://cmake.org/) to follow this installation guide.   
-
-
-### Linux
-
-Generally, both GCC and CMake are installed by default on Linux platforms. However, in some cases you may need to install them. For example, Ubuntu 18.04 comes with an older compiler that does not support some of the functionalities in this code directly. Therefore, you may want to [install a newer version of GCC](https://linuxize.com/post/how-to-install-gcc-compiler-on-ubuntu-18-04/). After this:
-
-1. Download the repository as a [*.zip file](https://github.com/battery-intelligence-lab/dtw-cpp/archive/refs/heads/main.zip), or clone it using following command: 
[…]
