+++ .github/workflows/ubuntu-unit.yml
@@ -25,9 +25,6 @@ jobs:
           - CC: gcc-12
             CXX: g++-12
             compiler: gcc-12 g++-12
-          - CC: gcc-13
-            CXX: g++-13
-            compiler: gcc-13 g++-13
           - CC: clang-12
             CXX: clang++-12
             compiler: clang-12 libomp-12-dev
