### 2024-08-16 6c1a14e
Linking OpenMP
