### 2024-08-16 Battery-Intelligence-Lab/DTWpp#16 @ElektrikAkar
Hi,

Thank you very much for your comment! We have fully updated the installation documentation https://github.com/Battery-Intelligence-Lab/dtw-cpp/pull/20.

- We integrated CMake installation into the bash commands.
- OpenMP is mostly included in GCC and MSVC installations; however, we included the necessary command for Apple Clang and added a note on installing the required OpenMP version for Ubuntu.
- For Gurobi, we linked to the [visual guidelines](https://battery-intelligence-lab.github.io/dtw-cpp/1_getting_started/1_installation.html#gurobi) provided by the company. However, since it is a somewhat lengthy process and an optional dependency, we chose not to copy and paste all the steps.
- Lastly, we integrated [CPMLicenses.cmake](https://github.com/cpm-cmake/CPMLicenses.cmake) to print the licenses of some dependencies included in the binary executable.

Please let us know if any further clarification on the third-party software is needed.
