+++ docs/1_getting_started/1_installation.md
@@ -1,6 +1,6 @@
 ---
 layout: default
-title: Running
+title: Installation
 nav_order: 2
 ---
 
