### 2024-08-16 Battery-Intelligence-Lab/DTWpp#12 @ElektrikAkar
Hi,

Thank you very much for both your comment! You are right that there was a lot of confusion in the installation section. Now, we have fully updated the installation documentation https://github.com/Battery-Intelligence-Lab/dtw-cpp/pull/20

- We changed order of the items.
- Corrected folder names.
- Navitaged users to `bin` folder by `cd ../bin` command after building the software inside `build` folder.
- Expanded the explanations for each operating system instead of referring to each other.
- Added an [example project](https://github.com/Battery-Intelligence-Lab/dtw-cpp/tree/main/examples/example_project) inside examples to import DTW-C++ as a library.
- Also corrected some of the default path variables defined inside CMake; where we seperated the `root` path as the path of the main project and `dtwc_folder` as the location of our library for finding example data easily.

Hopefully now it looks much better: https://battery-intelligence-lab.github.io/dtw-cpp/1_getting_started/1_installation.html

Please let us know if there are any further changes we can make to improve the quality of the documentation.
