### 2024-08-16 Battery-Intelligence-Lab/DTWpp#16 @i64
Hi @ElektrikAkar,

I checked the commit. LGTM!
