### 2024-08-16 #6881 @logological
@ElektrikAkar There appears to be a typo in your BibTeX file that breaks the hyperlink to tslearn.  In the file `paper.bib`, could you please change the line `   url = {https://github.com/tslearn-team/tslearn.},` to `   url = {https://github.com/tslearn-team/tslearn},`?  (That is, delete the period character at the end of the URL.)
