+++ joss/paper.bib
@@ -114,15 +114,17 @@ url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
 }
 
 @misc{Tavenard2020,
-   abstract = {tslearn is a general-purpose Python machine learning library for time series that offers tools for pre-processing and feature extraction as well as dedicated models for clustering, classification and regression. It follows scikit-learn's Application Programming Interface for transformers and estimators, allowing the use of standard pipelines and model selection tools on top of tslearn objects. It is distributed under the BSD-2-Clause license, and its source code is available at https://github.com/tslearn-team/tslearn.},
-   author = {Romain Tavenard and Johann Faouzi and Gilles Vandewiele and Felix Divo and Guillaume Androz and Chester Holtz and Marie Payne and Roman Yurchak and Marc Rußwurm},
-   journal = {Journal of Machine Learning Research},
-   keywords = {classification,clustering,data mining,pre-processing,time series},
-   pages = {1-6},
-   title = {Tslearn, A Machine Learning Toolkit for Time Series Data},
-   volume = {21},
-   url = {https://github.com/tslearn-team/tslearn.},
-   year = {2020},
+  author  = {Romain Tavenard and Johann Faouzi and Gilles Vandewiele and 
+             Felix Divo and Guillaume Androz and Chester Holtz and 
+             Marie Payne and Roman Yurchak and Marc Ru{\ss}wurm and 
+             Kushal Kolar and Eli Woods},
+  title   = {Tslearn, A Machine Learning Toolkit for Time Series Data},
+  journal = {Journal of Machine Learning Research},
+  year    = {2020},
+  volume  = {21},
+  number  = {118},
+  pages   = {1-6},
+  url     = {http://jmlr.org/papers/v21/20-091.html},
 }
 
 @article{Huangfu2018,
