### 2024-08-16 #6881 @logological
@ElektrikAkar I've spotted a few further issues with your references:

1. Sakoe & Chiba is missing the issue number (1).  You can add this to the BibTeX record with `number = 1,`.
2. Tavenard et al. (2020) is missing the issue number (118). You can add this to the BibTeX record with `number = 118,`.
3. You might consider having the `url` field for Tavenard et al. (2020) set to the location of the paper (`http://jmlr.org/papers/v21/20-091.html`) rather than the GitHub repository (`https://github.com/tslearn-team/tslearn`).
3. The URL supplied for the Gurobi Optimizer Reference Manual goes to the company's home page rather than the manual itself.  Consider using the exact URL for the manual: `https://www.gurobi.com/documentation/11.0/refman/`  (substitute `10.0`, `9.5`, etc. for `11.0` if you meant to cite an earlier version of the manual).
