### 2024-08-16 Battery-Intelligence-Lab/DTWpp#12 @ElektrikAkar
Hi, as both reviewers are happy with the current state, I am closing the issue. Many thanks for your efforts.
