+++ joss/paper.bib
@@ -23,6 +23,7 @@
    pages = {43-49},
    title = {Dynamic Programming Algorithm Optimization for Spoken Word Recognition},
    volume = {26},
+   number = {1},
    year = {1978},
 }
 
@@ -140,6 +141,7 @@ url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
    publisher = {Springer Verlag},
    title = {Parallelizing the dual revised simplex method},
    volume = {10},
+   number = {1},
    year = {2018},
 }
 
@@ -147,7 +149,7 @@ url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
   author = {{Gurobi Optimization, LLC}},
   title = {{Gurobi Optimizer Reference Manual}},
   year = 2024,
-  url = "https://www.gurobi.com"
+  url = "https://www.gurobi.com/documentation/11.0/refman/"
 }
 
 @misc{kumtepeli2023fast,
