+++ joss/paper.bib
@@ -87,19 +87,6 @@ month = {October},
 url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
 }
 
-@article{Sakoe1978,
-   abstract = {This paper reports on an optimum dynamic programming (DP) based time-normalization algorithm for spoken word recognition. First, a general principle of time-normalization is given using time-warping function. Then, two time-normalized distance definitions, called symmetric and asymmetric forms, are derived from the principle. These two forms are compared with each other through theoretical discussions and experimental studies. The symmetric form algorithm superiority is established. A new technique, called slope constraint, is successfully introduced, in which the warping function slope is restricted so as to improve discrimination between words in different categories. The effective slope constraint characteristic is qualitatively analyzed, and the optimum slope constraint condition is determined through experiments. The optimized algorithm is then extensively subjected to experimental comparison with various DP-algorithms, previously applied to spoken word recognition by different research groups. The experiment shows that the present algorithm gives no more than about two-thirds errors, even compared to the best conventional algorithm. © 1978 IEEE},
-   author = {Hiroaki Sakoe and Seibi Chiba},
-   doi = {10.1109/TASSP.1978.1163055},
-   issn = {00963518},
-   issue = {1},
-   journal = {IEEE Transactions on Acoustics, Speech, and Signal Processing},
-   pages = {43-49},
-   title = {Dynamic Programming Algorithm Optimization for Spoken Word Recognition},
-   volume = {26},
-   year = {1978},
-}
-
 @article{Rajabi2020,
    abstract = {Smart meters have been widely deployed in power networks since the last decade. This trend has resulted in an enormous volume of data being collected from the electricity customers. To gain benefits for various stakeholders in power systems, proper data mining techniques, such as clustering, need to be employed to extract the underlying patterns from energy consumptions. In this paper, a comparative study of different techniques for load pattern clustering is carried out. Different parameters of the methods that affect the clustering results are evaluated and the clustering algorithms are compared for two data sets. In addition, the two suitable and commonly used data size reduction techniques and feature definition/extraction methods for load pattern clustering are analysed. Furthermore, the existing studies on clustering of electricity customers are reviewed and the main results are highlighted. Finally, the future trends and major applications of clustering consumption patterns are outlined to inform industry practitioners and academic researchers to optimize smart meter operational use and effectiveness.},
    author = {Amin Rajabi and Mohsen Eskandari and Mojtaba Jabbari Ghadi and Li Li and Jiangfeng Zhang and Pierluigi Siano},
