### 2024-08-17 #6881 @logological
@ElektrikAkar Thanks, but it seems your edits have inadvertently introduced or revealed some further errors:

1. Huangfu & Hall (2018) now has both an `issue` and a `number` field, but only one of these should be specified.  (The way you have it now, LaTeX is typesetting the issue number as "(1, 1)".)  You can delete the `issue` field.
2. Sakoe & Chiba (1978) has the same problem as Huangfu & Hall (2018).  Please delete the `issue` field.
3. I hadn't noticed before that you had the wrong entry type for Tavenard et al. (2020), and as a result the formatting of the volume, issue, and page numbers differs from the other articles.  Please change its type from `@misc` to `@article`.

Could you please fix these?
