### 2024-08-17 cce8472
Keep only "issue" or "number" fields.
