+++ joss/paper.bib
@@ -18,7 +18,6 @@
    author = {Hiroaki Sakoe and Seibi Chiba},
    doi = {10.1109/TASSP.1978.1163055},
    issn = {00963518},
-   issue = {1},
    journal = {IEEE Transactions on Acoustics, Speech, and Signal Processing},
    pages = {43-49},
    title = {Dynamic Programming Algorithm Optimization for Spoken Word Recognition},
@@ -101,7 +100,7 @@ url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
    year = {2020},
 }
 
-@misc{Tavenard2020,
+@article{Tavenard2020,
   author  = {Romain Tavenard and Johann Faouzi and Gilles Vandewiele and 
              Felix Divo and Guillaume Androz and Chester Holtz and 
              Marie Payne and Roman Yurchak and Marc Ru{\ss}wurm and 
@@ -120,7 +119,6 @@ url = {https://www.cs.ucr.edu/~eamonn/time_series_data_2018/}
    author = {Q. Huangfu and J. A.J. Hall},
    doi = {10.1007/s12532-017-0130-5},
    issn = {18672957},
-   issue = {1},
    journal = {Mathematical Programming Computation},
    keywords = {Linear programming,Parallel computing,Revised simplex method},
    month = {3},
