### 2024-08-29 59511a6
version update from 1.0.0-rc0 to 1.0.0
