### 2024-08-29 #6881 @ElektrikAkar
Hi @logological,

Thank you very much for your patience. We have corrected ORCIDs, and released software on Github and Zenodo.

https://github.com/Battery-Intelligence-Lab/dtw-cpp/releases/tag/v1.0.0

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13551469.svg)](https://doi.org/10.5281/zenodo.13551469)

Hopefully everything is alright. Looking forward to your further instructions. Have a nice Friday and weekend.
