### 2024-08-31 #6881 @logological
@ElektrikAkar Thanks for confirming you've completed the remaining steps.  As far as I can tell, all the data is correct.  One thing that rather belatedly occurs to me is that the paper is on the long side for a JOSS article—about 2500 words, excluding equations, tables, and references—though the fact that neither I nor the reviewers noticed this earlier speaks to the relevance and usefulness of the entire text.  I'm happy to recommend publication as-is, though I may need to defer to the decision of the editor-in-chief in the event of a disagreement on this point.

Finally, thanks again to @ZhenchenHong and @i64 for your service as reviewers!
