### 2024-09-02 #6881 @ElektrikAkar
Hi @logological,

Thank you both, as well as the reviewers, for your efforts. We initially considered the word limit, but unfortunately, we must have lost track of it during the iterations to improve the manuscript. We are happy that you found the text useful and suggested it for acceptance as-is. Hopefully, the editor-in-chief will also agree. In any case, please let us know if there are any further steps we should take. Many thanks, and have a nice week ahead.
