### 2024-09-06 #6881 @ElektrikAkar
Hi @crvernon

Thank you very much for your comment! Yes, I had forgotten this and just edited. Now Zenodo title should be same as the JOSS paper title.

https://zenodo.org/records/13551469
