### 2024-09-07 #6881 @ElektrikAkar
We thank you, editors, reviewers for their effort to make the manuscript better. The review process was great, it was a pleasure to meet you. We will forward to our next publication and/or contributing to JOSS via being a reviewer. Have a nice weekend :blush:
